+++ src/containers/Activity/ActivityCardList/CardFooter/FooterExtra/NotMember/index.js
@@ -13,7 +13,7 @@ const NotMember = ({
       {hasforbiddened ?
         <Flex justify='center'>
           <div className='activityCard__container-forbiddened'>
-            {reportIcon} 被举报次数过多，已扣分
+            {reportIcon} 举报过多，已扣分
           </div>
         </Flex>
         :
