+++ src/containers/Activity/ActivityCardList/CardHeader/cardHeader.css
@@ -22,7 +22,6 @@
   font-stretch: normal;
   line-height: 1;
   letter-spacing: normal;
-  text-align: center;
   color: #898fa3;
   margin-top: 0.52vh;
 }
