+++ src/utils/ConvertTsToDate/index.js
@@ -1,0 +1,14 @@
+const convertTsToDate = (timestamp) => {
+  const date = new Date(timestamp)
+  const year = date.getFullYear()
+  const month = date.getMonth() + 1
+  const day = date.getDate() < 10 ? `0${date.getDate()}` : date.getDate()
+  const hour = date.getHours() < 10 ? `0${date.getHours()}` : date.getHours()
+  const minute = date.getMinutes() < 10 ? `0${date.getMinutes()}` : date.getMinutes()
+  const second = date.getSeconds() < 10 ? `0${date.getSeconds()}` : date.getSeconds()
+  return `${year}.${month}.${day} ${hour}:${minute}:${second}`
+}
+
+export {
+  convertTsToDate,
+}
