+++ src/utils/ErrorCode/index.js
@@ -1,0 +1,8 @@
+const errorCode = {
+  413: '图片大于512KB O(∩_∩)O~',
+  403: '上传的不是图片哦 O(∩_∩)O~~'
+}
+
+export {
+  errorCode,
+}
