+++ src/utils/constructPicUrl/index.js
@@ -1,0 +1,9 @@
+import { config } from '../../config'
+
+const constructPicUrl = (name) => {
+  return `${config.prefixUrl}${name}${config.imgSlim}`
+}
+
+export {
+  constructPicUrl,
+}
